+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Changed: `cargo mutants` now tries to match the behavior of `cargo test` when run within a workspace. If run in a package directory, it tests only that package. If run in a workspace that is not a package (a "virtual workspace"), it tests the configured default packages, or otherwise all packages. This can all be overridden with the `--package` or `--workspace` options.
+
 - New: generate key-value map values from types like `BTreeMap<String, Vec<u8>>`.
 
 - Changed: Send trace messages to stderr rather stdout, in part so that it won't pollute json output.